merge with pace.

+++ include/benchmark.h
@@ -32,7 +32,7 @@ fp_t
 stdev(const std::vector<fp_t>&);
 
 void
-b_decoder_ler(Decoder*, uint32_t shots, std::mt19937_64&,
+b_decoder_ler(Decoder*, uint64_t shots, std::mt19937_64&,
         bool save_per_shot_data=true);
 
 }  // qrc
